Major backend rework, now uses SQLite for most backing storage

+++ src/embeds/include/MatchEmbeds.h
@@ -22,7 +22,7 @@ public:
     static dpp::embed matchNotFound(int matchID);
     static dpp::embed matchAlreadyPlayed(int matchID);
     static dpp::embed matchPlayersNotRegistered(vector<string> unregistered);
-    static dpp::embed matchPlayersNotOnTeam(vector<unsigned long long> teamless, int matchID);
+    static dpp::embed matchPlayersNotOnTeam(vector<int64_t> teamless, int matchID);
 };
